:recycle: Improved the files screen
